requirements.txt
python-telegram-bot[all]==21.0.1
aiosqlite==0.19.0